+++ legacy/edje/src/lib/edje_text.c
@@ -118,7 +118,7 @@ _edje_text_fit_x(Edje *ed, Edje_Real_Part *ep,
 
    sc = ed->scale;
    if (sc == ZERO) sc = _edje_scale;
-   
+
    *free_text = 0;
    if (sw <= 1) return "";
 
@@ -161,7 +161,7 @@ _edje_text_fit_x(Edje *ed, Edje_Real_Part *ep,
     * FIXME: we might want to set a max string length somewhere...
     */
    extra = 1 + 3 + 3; /* terminator, leading and trailing ellipsis */
-   orig_len = MIN(orig_len, 8192 - extra);
+   orig_len = MIN(orig_len, ((size_t) 8192 - extra));
 
    if (!(buf = malloc(orig_len + extra)))
      return text;
@@ -216,7 +216,7 @@ _edje_text_fit_x(Edje *ed, Edje_Real_Part *ep,
 		  break;
 	       }
 	  }
-	else if ((c1 > 0 && c1 >= orig_len) || c2 == 0)
+	else if ((c1 > 0 && (size_t) c1 >= orig_len) || c2 == 0)
 	  {
 	     buf[0] = 0;
 	     break;
@@ -256,7 +256,7 @@ _edje_text_font_get(const char *base, const char *new, char **free_later)
 
    font_len = strlen(new);
    aux = strchr(base_style, ',');
-   style_len = (aux) ?  (aux - base_style) : strlen(base_style);
+   style_len = (aux) ? (aux - base_style) : (int) strlen(base_style);
 
    *free_later = malloc(font_len + style_len + 1);
    memcpy(*free_later, new, font_len);
